requirements.txt
django==4.2.5
psycopg2_binary==2.9.7
djangorestframework==3.14.0
django-filter==23.2
Pillow==10.0.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
pytest-django==4.5.2
django-cors-headers==4.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
drf-yasg==1.21.7